# api.yaml
openapi: 3.0.0
info:
  title: Personal-Trainer-API
  version: 1.0.0
servers:
  - url: /api/v1
    description: Local development
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    BaseResponse:
      type: object
      required: [status, message, code]
      properties:
        status:
          type: string
          enum: [success, error]
        message:
          type: string
        code:
          type: string
          description: Machine-readable response code (e.g., OK, BAD_REQUEST, NOT_FOUND)
    SuccessResponse:
      allOf:
        - $ref: "#/components/schemas/BaseResponse"
        - type: object
          properties:
            data:
              nullable: true
              description: Any JSON value (object/array/string/number/etc)
              x-go-type: interface{}
            meta:
              nullable: true
              description: Any JSON value (usually object)
              x-go-type: interface{}
    ErrorResponse:
      allOf:
        - $ref: "#/components/schemas/BaseResponse"
        - type: object
          properties:
            errors:
              type: array
              nullable: true
              items:
                $ref: "#/components/schemas/FieldError"
    FieldError:
      type: object
      required: [field, message]
      properties:
        field:
          type: string
          description: The field that failed validation
        message:
          type: string
          description: Human-readable error message for this field
...
    UpdateProfileRequest:
      type: object
      description: |
        Partial profile update. avatar_url is intentionally NOT on this
        request — avatars are set exclusively via
        POST /users/me/profile/picture (which uploads through MinIO and
        writes the column asynchronously). Putting avatar_url here too
        would let the JSON endpoint clobber a freshly-uploaded picture
        and was previously also forcing profile_complete = false until
        a picture existed.
      properties:
        name:
          type: string
          example: "Ezeaku Dorothy"
        gender:
          type: string
          enum: [male, female, other]
          example: female
        fitness_goals:
          type: array
          items:
            type: string
            enum:
              - lose_weight
              - build_muscle
              - improve_flexibility
              - boost_energy
              - build_healthy_habits
              - build_strength
          minItems: 1
          example: [build_muscle, build_strength]
        fitness_level:
          type: string
          enum: [beginner, intermediate, advanced]
          example: beginner
    UserProfileResponse:
      type: object
      required:
        [
          id,
          email,
          name,
          gender,
          fitness_goals,
          fitness_level,
          avatar_url,
          profile_complete,
        ]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        gender:
          type: string
          nullable: true
        fitness_goals:
          type: array
          items:
            type: string
        fitness_level:
          type: string
          nullable: true
        avatar_url:
          type: string
          nullable: true
        profile_complete:
          type: boolean
...
    RescheduleBookingRequest:
      type: object
      required:
        - new_datetime
        - timezone
        - reason
      properties:
        new_datetime:
          type: string
          format: date-time
          description: The new desired date and time (ISO 8601)
        timezone:
          type: string
          description: IANA timezone string (e.g. America/New_York)
          example: Africa/Lagos
        reason:
          $ref: "#/components/schemas/RescheduleReason"
        notes:
          type: string
          maxLength: 200
          description: Optional additional details
        phone_number:
          type: string
          description: Updated phone number (phone_callback bookings only)
    RescheduleReason:
      type: string
      enum:
        - something_came_up
        - feeling_unwell
        - work_conflict
        - personal_emergency
        - travel
        - other
...
    BookDiscoveryCallRequest:
      type: object
      required: [name, email, contact_mode, selected_datetime, timezone]
      properties:
        name:
          type: string
          example: Jane Doe
        email:
          type: string
          format: email
          example: jane@example.com
        contact_mode:
          type: string
          enum: [zoom_meeting, phone_callback]
        phone_number:
          type: string
          description: Required when contact_mode is phone_callback
          example: "+2348012345678"
        selected_datetime:
          type: string
          format: date-time
          description: ISO 8601 UTC datetime for the call
          example: "2026-05-20T10:00:00Z"
        timezone:
          type: string
          description: Client IANA timezone
          example: America/New_York
    BookingSlotRequest:
      type: object
      required: [day_of_week, start_time, end_time]
      properties:
        day_of_week:
          type: integer
          minimum: 0
          maximum: 6
          description: 0=Sunday, 6=Saturday
          example: 1
        start_time:
          type: string
          description: 24h time HH:MM
          example: "09:00"
        end_time:
          type: string
          description: 24h time HH:MM
          example: "17:00"
        timezone:
          type: string
          description: IANA timezone for the slot
          example: Africa/Lagos
        is_active:
          type: boolean
          default: true
...
security:
  - bearerAuth: []
paths:
...
  /users/me/profile:
    patch:
      operationId: UpdateUserProfile
      tags:
        - User
      summary: Update the authenticated user's profile (onboarding)
      description: |
        Saves name, gender, fitness goals, and fitness level for the logged-in
        user. All fields are optional — send only what changed. avatar_url is
        NOT accepted here; upload via POST /users/me/profile/picture.
        profile_complete flips to true once name, gender, and fitness_level
        are all set.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
      responses:
        "200":
          description: Profile updated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/UserProfileResponse"
              example:
                status: success
                message: Profile updated successfully
                code: OK
                data:
                  id: "550e8400-e29b-41d4-a716-446655440000"
                  email: "dorothy@example.com"
                  name: "Ezeaku Dorothy"
                  gender: "female"
                  fitness_goals: ["build_muscle", "build_strength"]
                  fitness_level: "beginner"
                  avatar_url: null
                  profile_complete: true
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                status: error
                message: Missing or invalid access token
                code: UNAUTHORIZED
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    get:
      operationId: GetUserProfile
      tags:
        - User
      summary: Get the authenticated user's profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Profile fetched
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/UserProfileResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /bookings/upcoming:
    get:
      operationId: GetUpcomingBookings
      tags:
        - Booking
      summary: Get upcoming bookings for the authenticated client
      description: |
        Returns a paginated list of all upcoming bookings for the authenticated client.
        Includes both paid training sessions and discovery calls, sorted by date ascending.
        Times are returned in the client's local timezone when a timezone is provided.
      security:
        - bearerAuth: []
      parameters:
        - name: timezone
          in: query
          required: false
          schema:
            type: string
            example: Africa/Lagos
          description: IANA timezone for displaying times (e.g. America/New_York)
        - name: type
          in: query
          required: false
          schema:
            type: string
            enum: [paid_session, discovery_call]
          description: Filter by booking type
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
      responses:
        "200":
          description: Upcoming bookings retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Bad request — invalid query parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /bookings/discovery:
    post:
      security:
        - bearerAuth: []
      tags:
        - Booking
      operationId: BookDiscoveryCall
      summary: Book a discovery call with a FitCall rep
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookDiscoveryCallRequest"
      responses:
        "201":
          description: Discovery call booked successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Time slot already booked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /bookings/{id}/reschedule:
    put:
      operationId: RescheduleDiscoveryCall
      tags:
        - Booking
      summary: Reschedule an existing discovery call booking
      description: |
        Allows a client to reschedule their discovery call to a new date and time.
        Blocked within 12 hours of the original call time.
        For Zoom bookings, the old meeting is deleted and a new one is created.
        For phone callbacks, only the time and phone number are updated.
        Maximum 3 reschedules per booking.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RescheduleBookingRequest"
      responses:
        "200":
          description: Discovery call rescheduled successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Invalid input or new time outside open hours / in the past
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Booking locked (within 12-hour window), cancelled, or completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Booking not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: New time slot already taken
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Maximum reschedule limit (3) reached
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /discovery-slots:
    get:
      security: []
      operationId: GetDiscoverySlots
      tags:
        - Booking
      summary: List all active discovery slots (public)
      parameters:
        - name: timezone
          in: query
          required: false
          schema:
            type: string
          description: IANA timezone to convert slots into (e.g. America/New_York)
      responses:
        "200":
          description: Booking slots retrieved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      security:
        - bearerAuth: []
      operationId: CreateDiscoverySlot
      summary: Create a discovery slot (admin or customer_care only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingSlotRequest"
      responses:
        "201":
          description: Booking slot created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /discovery-slots/{id}:
    put:
      security:
        - bearerAuth: []
      tags:
        - Booking
      operationId: UpdateDiscoverySlot
      summary: Update a discovery slot (admin or customer_care only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingSlotRequest"
      responses:
        "200":
          description: Booking slot updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Slot not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      security:
        - bearerAuth: []
      operationId: DeleteDiscoverySlot
      summary: Delete a discovery slot (admin or customer_care only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Booking slot deleted
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Slot not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  # Get booking slot by trainer id
...